test: improve scenario for E2E test
